Create TraversalItem class
TraversalItem is an item in a traversal. It only holds a reference to an item from the virtual graph (VirtualVertex | VirtualEdge), and a reference to the previous traversal item in the traversal. This is needed to be able to gain access to the traversal path, and to use utilities like otherV.